blockchainer/network: move StateSync interface to the user
Only network package cares about it.
