+++ pkg/network/state_sync.go
@@ -1,6 +1,7 @@
-package blockchainer
+package network
 
 import (
+	"github.com/nspcc-dev/neo-go/pkg/core/blockchainer"
 	"github.com/nspcc-dev/neo-go/pkg/core/mpt"
 	"github.com/nspcc-dev/neo-go/pkg/util"
 )
@@ -8,7 +9,7 @@ import (
 // StateSync represents state sync module.
 type StateSync interface {
 	AddMPTNodes([][]byte) error
-	Blockqueuer // Blockqueuer interface
+	blockchainer.Blockqueuer // Blockqueuer interface
 	Init(currChainHeight uint32) error
 	IsActive() bool
 	IsInitialized() bool
