+++ pkg/rpc/server/server_helper_test.go
@@ -102,7 +102,7 @@ func initClearServerWithServices(t testing.TB, needOracle bool, needNotary bool)
 
 	serverConfig := network.NewServerConfig(cfg)
 	serverConfig.Port = 0
-	server, err := network.NewServer(serverConfig, chain, logger)
+	server, err := network.NewServer(serverConfig, chain, chain.GetStateSyncModule(), logger)
 	require.NoError(t, err)
 	rpcServer := New(chain, cfg.ApplicationConfiguration.RPC, server, orc, logger)
 	errCh := make(chan error, 2)
